+++ src/main/java/forestry/api/apiculture/ForestryBeeSpecies.java
@@ -59,5 +59,12 @@ public class ForestryBeeSpecies {
 	public static final ResourceLocation LUSH = forestry("bee_lush");
 	public static final ResourceLocation VERDANT = forestry("bee_verdant");
 	public static final ResourceLocation LUXURIANT = forestry("bee_luxuriant");
+	public static final ResourceLocation KLEPTOPLASTIC = forestry("bee_kleptoplastic");
+	public static final ResourceLocation PHOTOSYNTHETIC = forestry("bee_photosynthetic");
+	public static final ResourceLocation AUTOTROPHIC = forestry("bee_autotrophic");
+	public static final ResourceLocation AQUATIC = forestry("bee_aquatic");
+	public static final ResourceLocation PIRATE = forestry("bee_pirate");
+	public static final ResourceLocation PRISMATIC = forestry("bee_prismatic");
+	public static final ResourceLocation ABYSSAL = forestry("bee_abyssal");
 	public static final ResourceLocation SHULKING = forestry("bee_shulking");
 }