+++ src/main/java/forestry/apiculture/items/EnumHoneyComb.java
@@ -32,7 +32,9 @@ public enum EnumHoneyComb implements StringRepresentable, IItemSubtype, IBlockSu
 	WHEATEN(new Color(0xfeff8f), new Color(0xffffff)),
 	MOSSY(new Color(0x2a3313), new Color(0x7e9939)),
 	MELLOW(new Color(0x886000), new Color(0xfff960)),
-	KAOLIN(new Color(0x5e6c8d), new Color(0xafb9d6));
+	KAOLIN(new Color(0x5e6c8d), new Color(0xafb9d6)),
+	SPONGE(new Color(0x9D8F39), new Color(0xe1e351)),
+	;
 	//""(new Color(0xd7bee5), new Color(0xfd58ab)); // kindof pinkish
 
 	public static final EnumHoneyComb[] VALUES = values();